+++ fast_barcode_scanner/lib/src/overlays/code_boundary_overlay/code_border_painter.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 
 import '../../../fast_barcode_scanner.dart';
 
-class CodeBoarderPainter extends CustomPainter {
+class CodeBorderPainter extends CustomPainter {
   final CodeBorderPaintBuilder? barcodePaintSelector;
   final CodeValueDisplayBuilder? textDecorator;
 
-  CodeBoarderPainter({
+  CodeBorderPainter({
     required this.imageSize,
     required this.barcodes,
     this.barcodePaintSelector,
@@ -16,7 +16,7 @@ class CodeBoarderPainter extends CustomPainter {
   final Size imageSize;
   final List<Barcode> barcodes;
 
-  final _standardPaint = Paint()
+  static final _standardPaint = Paint()
     ..style = PaintingStyle.stroke
     ..strokeWidth = 2.0
     ..color = Colors.red;
@@ -99,7 +99,7 @@ class CodeBoarderPainter extends CustomPainter {
   }
 
   @override
-  bool shouldRepaint(CodeBoarderPainter oldDelegate) {
+  bool shouldRepaint(CodeBorderPainter oldDelegate) {
     return oldDelegate.imageSize != imageSize ||
         oldDelegate.barcodes != barcodes;
   }
