+++ fast_barcode_scanner/lib/src/overlays/code_boundary_overlay/code_boundary_overlay.dart
@@ -30,7 +30,7 @@ class _CodeBoundaryOverlayState extends State<CodeBoundaryOverlay> {
           final analysisSize = _cameraController.analysisSize;
           if (analysisSize != null && barcodes.isNotEmpty) {
             return CustomPaint(
-              painter: CodeBoarderPainter(
+              painter: CodeBorderPainter(
                 imageSize: analysisSize,
                 barcodes: barcodes,
                 barcodePaintSelector: widget.codeBorderPaintBuilder,
